Move definitions and MavServer to their own files
