+++ gzsitl/CMakeLists.txt
@@ -35,7 +35,10 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${GAZEBO_CXX_FLAGS}")
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${CFLAGS} ${CUSTOM_COMPILE_FLAGS}")
 
 # Build Plugin
-add_library(gzsitl_plugin SHARED gzsitl_plugin.cc)
+add_library(gzsitl_plugin SHARED
+    gzsitl_plugin.cc
+    mavserver.cc
+)
 target_link_libraries(gzsitl_plugin ${GAZEBO_libraries} ${LDFLAGS})
 
 # Set Gazebo Environment Variables
